kernel: define off_t

+++ moose/fs/fat.c
@@ -226,7 +226,7 @@ static int pfatfs__write(pfatfs *fs, const void *buf, size_t size) {
     return 0;
 }
 
-static int pfatfs__seek(pfatfs *fs, i32 off, int whence) {
+static int pfatfs__seek(pfatfs *fs, off_t off, int whence) {
     ssize_t result = fs->settings->seek(fs->settings->handle, off, whence);
     if (result != 0)
         return -EIO;
@@ -239,14 +239,14 @@ static u32 pfatfs__cluster_to_bytes(pfatfs *fs, u32 cluster, u16 offset) {
     return (cluster - 2) * fs->bytes_per_cluster + offset + fs->data_offset;
 }
 
-static int pfatfs__seek_read(pfatfs *fs, i32 off, void *buf, size_t size) {
+static int pfatfs__seek_read(pfatfs *fs, off_t off, void *buf, size_t size) {
     int result = pfatfs__seek(fs, off, SEEK_SET);
     if (result == 0)
         result = pfatfs__read(fs, buf, size);
     return result;
 }
 
-static int pfatfs__seek_write(pfatfs *fs, i32 off, const void *buf,
+static int pfatfs__seek_write(pfatfs *fs, off_t off, const void *buf,
                               size_t size) {
     int result = pfatfs__seek(fs, off, SEEK_SET);
     if (result == 0)
@@ -904,7 +904,7 @@ int pfatfs_truncate(pfatfs *fs, pfatfs_file *file, size_t length) {
     return 0;
 }
 
-int pfatfs_seek(pfatfs *fs, pfatfs_file *file, ssize_t offset, int whence) {
+int pfatfs_seek(pfatfs *fs, pfatfs_file *file, off_t offset, int whence) {
     u32 new_offset;
     switch (whence) {
     case SEEK_CUR:
